fixed extra quizzer
